refactor(plugins): plugin命名标准化
