+++ packages/plugin/src/index.ts
@@ -1,9 +1,9 @@
-import TimeBar from './timeBar';
-import SnapLine from './snapLine';
+import TimeBar from './time-bar';
+import SnapLine from './snap-line';
 import Tooltip from './tooltip';
 import Legend from './legend';
-import ZoomSlider from './zoomSlider';
-import EdgeFilterLens from './edgeFilterLens';
+import ZoomSlider from './zoom-slider';
+import EdgeFilterLens from './edge-filter-lens';
 import Fisheye from './fisheye';
 import MiniMap from './minimap';
 import Bundling from './bundling';
