+++ packages/plugin/src/time-bar/dispatcher.ts
