+++ packages/plugin/src/time-bar/type/control-bar.ts
