+++ packages/plugin/src/time-bar/type/slider.ts
