+++ packages/plugin/src/time-bar/ui/control-bar/control-btn.ts
