+++ packages/plugin/src/time-bar/ui/control-bar/play-btn.ts
