+++ packages/plugin/src/time-bar/ui/progress-bar/index.ts
