+++ packages/plugin/src/time-bar/ui/progress-bar/tick.ts
