+++ packages/plugin/src/time-bar/ui/slider/simple-slider.ts
