+++ packages/plugin/src/time-bar/ui/slider/slider-logic.ts
