+++ packages/plugin/src/time-bar/ui/slider/trend-slider.ts
