+++ packages/plugin/src/time-bar/utils/path.ts
