+++ packages/plugin/src/time-bar/utils/player.ts
