+++ packages/plugin/src/zoom-slider/index.ts
